refactor: file structure refactored
