+++ Assignment_4_ExpenseTracker/Constants/Enumerations.cs
@@ -1,0 +1,41 @@
+﻿namespace Constants
+{
+    namespace Enumerations
+    {
+        public enum ExpenseOptions
+        {
+            Food = 1,
+            Travel,
+            Electricity,
+            Stay,
+            Education,
+            Clothing,
+            Gadgets,
+            Grocery,
+            Invest,
+            Grooming,
+            Other
+        }
+
+        public enum IncomeOptions
+        {
+            Salary = 1,
+            FreeLancing,
+            Loan,
+            Other
+        }
+
+        public enum MainMenu
+        {
+            Add_Income = 1,
+            Add_Expense,
+            Close_App
+        }
+
+        public enum SetDateOptions
+        {
+            Current_Date = 1,
+            Own_Date
+        }
+    }
+}
