--- Assignment_4_ExpenseTracker/Constants/Enumerations/ExpenseOptions.cs
@@ -1,20 +1,0 @@
-﻿namespace Constants
-{
-    namespace Enumerations
-    {
-        public enum ExpenseOptions
-        {
-            Food = 1,
-            Travel,
-            Electricity,
-            Stay,
-            Education,
-            Clothing,
-            Gadgets,
-            Grocery,
-            Invest,
-            Grooming,
-            Other
-        }
-    }
-}
