--- Assignment_4_ExpenseTracker/Constants/Enumerations/IncomeOptions.cs
@@ -1,13 +1,0 @@
-﻿namespace Constants
-{
-    namespace Enumerations
-    {
-        public enum IncomeOptions
-        {
-            Salary = 1,
-            FreeLancing,
-            Loan,
-            Other
-        }
-    }
-}
