--- Assignment_4_ExpenseTracker/Constants/Enumerations/MainMenu.cs
@@ -1,12 +1,0 @@
-﻿namespace Constants
-{
-    namespace Enumerations
-    {
-        public enum MainMenu
-        {
-            Add_Income = 1,
-            Add_Expense,
-            Close_App
-        }
-    }
-}
