--- Assignment_4_ExpenseTracker/Constants/Enumerations/SetTimeOptions.cs
@@ -1,11 +1,0 @@
-﻿namespace Constants
-{
-    namespace Enumerations
-    {
-        public enum SetTimeOptions
-        {
-            Current_Time = 1,
-            Own_Time
-        }
-    }
-}
