+++ Assignment_4_ExpenseTracker/Repository/FinanceRepository.cs
@@ -4,12 +4,11 @@ namespace Repository
 {
     public class FinanceRepository
     {
-        public List<IFinance> FinanceData { get; set; };
+        public List<IFinance> FinanceData { get; set; }
 
         public FinanceRepository()
         {
             FinanceData = new List<IFinance>();
         }
-
     }
 }
