+++ Assignment_4_ExpenseTracker/RepositoryManage/UpdateRepository.cs
@@ -4,7 +4,7 @@ using Models;
 using Assignment_4_ExpenseTracker.HelperUtility;
 using Constants.Enumerations;
 
-namespace Assignment_4_ExpenseTracker.RepositoryManager
+namespace Assignment_4_ExpenseTracker.RepositoryManage
 {
     public static class UpdateRepository
     {
@@ -13,7 +13,7 @@ namespace Assignment_4_ExpenseTracker.RepositoryManager
             (IncomeOptions, string?) incomeSource = GetUserData.GetIncomeSource();
             int amount = GetUserData.GetAmount();
             int actionId = IdGenerator.TransactionIdGenerator(financeData);
-            DateOnly actionDate = GetUserData.GetActivityTime();
+            DateOnly actionDate = GetUserData.GetActivityDate();
             financeData.Add(new Income(incomeSource.Item1, incomeSource.Item2, amount, actionId, actionDate));
         }
         public static void AddExpense(List<IFinance> financeData)
@@ -21,7 +21,7 @@ namespace Assignment_4_ExpenseTracker.RepositoryManager
             (ExpenseOptions, string?) expenseSource = GetUserData.GetExpenseSource();
             int amount = GetUserData.GetAmount();
             int actionId = IdGenerator.TransactionIdGenerator(financeData);
-            DateOnly actionDate = GetUserData.GetActivityTime();
+            DateOnly actionDate = GetUserData.GetActivityDate();
             financeData.Add(new Expense(expenseSource.Item1, expenseSource.Item2, amount, actionId, actionDate));
         }
     }
